extract find_cvar() to engine.rs
